% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_IPTW.R
\name{add_IPTW}
\alias{add_IPTW}
\title{Add an iptw column to a dataframe for a specific variable used as grouping factor}
\usage{
add_IPTW(data, tx_name, tx_levels = NULL, iptw = "iptw", ps_prefix = "PS_")
}
\arguments{
\item{data:}{the data.frame object for the analysis, output of add_PS()}

\item{tx_name:}{the column name of the treatment variable}

\item{tx_levels:}{the levels of the treatment variable (default: levels() of the column, it will be converted to factor() if is not)}

\item{iptw:}{the name of the added column with IPTW values}

\item{ps_prefix:}{the prefix for PS columns generated by add_PS() (default: "PS_")}
}
\value{
A data.frame object which is the original data.frame with a PSassign and IPTW column added
}
\description{
This function add Inverse Probability of Treatment Weighting (IPTW) columns
to the input data.frame object, you get a "PSassign" column of the actual PS
for the treatment indicated and its relative iptw as column too
}
