% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_PS.R
\name{add_PS}
\alias{add_PS}
\title{Add propensity scores to a dataframe for a specific grouping variable}
\usage{
add_PS(
  data,
  formula,
  prefix = "PS_",
  family = VGAM::multinomial(parallel = FALSE)
)
}
\arguments{
\item{data:}{the data.frame object for the analysis}

\item{formula:}{the regression model formula}

\item{prefix:}{the propensity score columns prefix to use (default: "PS_")}

\item{family:}{the family of regression to be used in vglm (default: "multinomial")}
}
\value{
A data.frame object which is the original data.frame passed with a PS column added
}
\description{
This function add propensity score (PS) columns to the input data.frame object
you get a PS column for each factor of the passed outcome in the formula
with a specified prefix (default "PS_")
}
