% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/balance_cov.R
\name{balance_cov}
\alias{balance_cov}
\title{Calculate unadjusted and IPTW-adjusted Standardised Mean Difference (SMD) for each variable based on the grouping factor}
\usage{
balance_cov(
  data,
  db_iptw,
  covariates,
  outcome,
  round_factor = 3,
  print = FALSE
)
}
\arguments{
\item{data:}{the data.frame object for the analysis}

\item{db_iptw:}{the data.frame object for the analysis, output of normalize_IPTW()}

\item{covariates:}{list of variables to be compared}

\item{outcome:}{grouping variable}
}
\value{
A list with unadjusted and adjusted comparison of variables based on outcome group and the design for weighted analysis (weighted)
}
\description{
This function calculates SMD for both the unadjusted and the IPTW-adjusted model
returning both models variable SMD and the model design used for adjusted (weighted)
}
