% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iptw_plot.R
\name{iptw_plot}
\alias{iptw_plot}
\title{Plot IPTW correction effect on Standardised Mean Difference (SMD) of variables passed as covariates}
\usage{
iptw_plot(unadjusted, adjusted, font_family = "sans-serif", font_size = 18)
}
\arguments{
\item{unadjusted:}{the unadjusted SMD stratification (output of balance_cov())}

\item{adjusted:}{the adjusted SMD stratification (output of balance_cov())}

\item{font_family:}{the family to use as plot font}

\item{font_size:}{the size to use as plot font}
}
\value{
A ggplot2 object that represent effect of IPTW normalization on SMD for each covariate
}
\description{
This function plots the result of Standardised Mean Difference (SMD) reduction
after adjustment for normalized IPTW. Cut-off 0.1 is highligted, unadjusted and
adjusted variable SMD are plotted as dotted and continuous lines respectively.
}
