% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize_IPTW.R
\name{normalize_IPTW}
\alias{normalize_IPTW}
\title{Add weights based on IPTW normalization to a dataframe}
\usage{
normalize_IPTW(
  data,
  iptw = "iptw",
  tx_name,
  tx_levels = NULL,
  round_factor = 2,
  print = FALSE
)
}
\arguments{
\item{data:}{the data.frame object for the analysis, output of add_IPTW()}

\item{iptw:}{the column name of the IPTW variable (output of add_IPTW())}

\item{tx_name:}{the column name of the treatment variable}

\item{tx_levels:}{the levels of the treatment variable (default: levels() of the column, it will be converted to factor() if is not)}

\item{round_factor:}{number of decimal inputs to round the stdout (default: 2)}

\item{print:}{resumen of mean comparison to stdout? (default: FALSE)}
}
\value{
A data.frame object which is the original data.frame with a 'sw' column added with weights based on IPTW
}
\description{
This function add an 'sw' column to the input data.frame containing weights
calculated on normalized IPTW on the number of samples within each group
}
