% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_covariates.R
\name{numeric_variable_summary}
\alias{numeric_variable_summary}
\title{Check all passed variables are in dataframe and type (factor vs numeric)}
\usage{
numeric_variable_summary(NUMERIC_VARIABLE_VECTOR)
}
\arguments{
\item{data:}{the data.frame input onject}

\item{covariates:}{the vector of variables to be analyzed}

\item{stdout:}{print summaries for each variable passed (default: TRUE)}

\item{convert:}{automatically converts characters and numeric with 0 and 1 to factors (default: FALSE)}
}
\value{
A stdout value of each variable type (numeric vs. factor)
}
\description{
This function check that all passed variables are present in data.frame
colnames and return a stdout of type of variable too (if requested)
}
