% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/riptw.R
\name{riptw}
\alias{riptw}
\title{Calculate normalized Inverse Probability of Treatment Weighting (IPTW) weights based on a set of covariates and the treatment variable specified}
\usage{
riptw(
  data,
  outcome,
  formula = NULL,
  outcome_levels = NULL,
  covariates,
  rcs_covariate = NULL,
  rcs_df = 5,
  ps_prefix = "PS_",
  stdout = FALSE,
  stdout_cov = TRUE,
  model_family = VGAM::multinomial(parallel = FALSE),
  iptw = "iptw",
  round_norm = 2,
  round_bc = 3,
  plot_font_family = "sans-serif",
  plot_font_size = 18
)
}
\arguments{
\item{data:}{the data.frame object for the analysis}

\item{outcome:}{the column name of the treatment variable}

\item{formula:}{the regression model formula as.formula() object (optional, you can pass outcome and covariates as vector)}

\item{outcome_levels:}{the levels of the treatment variable (optional: levels() of the column, it will be converted to factor() in case is necessary)}

\item{covariates:}{vector of model covariate names}

\item{rcs_covariate:}{the variable(s) to be used as restricted cubic spline (optional)}

\item{rcs_df:}{if rcs_covariate is specified the degrees of freedom to use for RCS transformation (optional, default: 5)}

\item{ps_prefix:}{the propensity score columns prefix to use for the output data.frame (default: "PS_")}

\item{stdout:}{print summaries for each function to stdout (default: FALSE)}

\item{stdout_cov:}{print summaries for each variable passed (default: TRUE)}

\item{model_family:}{the family of regression to be used in vglm (default: "multinomial")}

\item{round_norm:}{if stdout = TRUE selected, the number of decimal inputs to round the normalize_IPTW() stdout (default: 2)}

\item{round_bc:}{if stdout = TRUE selected, the number of decimal inputs to round the balance_cov() stdout (default: 3)}

\item{iptw:}{the name of the added column with IPTW values in output data.frame (default: "iptw")}

\item{plot_font_family:}{the family to use as output plot font (default: sans-serif)}

\item{plot_font_size:}{the size to use as output plot font (default: 18)}
}
\value{
A list object with 'data': the original data.frame with PS, IPTW, and weights columns added for each sample - 'plot': the SMD reduction plot - 'unadjusted': the covariates stratification table (before IPTW-adjustment) - 'adjusted': the covariates stratification table (after IPTW-adjustment)
}
\description{
This function calculates normalized Inverse Probability of Treatment Weighting (IPTW)
for each sample and add propensity score column (for each tratment variable group, "PS_" prefix),
IPTW ("iptw"), and normalized-IPTW ("sw") column to the original input data.frame.
It allows for both model formula direct specification or simply passing outcome and
covariates as a vector. It also allows for Restricted Cubic Spline (RCS) transformation
of some covariates. It also output unadjusted vs. adjusted Standardised Mean Differenc (SMD)
plot to compare SMD reduction for each covariate using IPTW, and unadjusted and adjusted
stratification tables for each covariate with the treatment variable groups (with p-values).
}
