% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_formula.R
\name{split_formula}
\alias{split_formula}
\title{Split passed formula into outcome and covariates}
\usage{
split_formula(formula)
}
\arguments{
\item{formula:}{the model formula passed}
}
\value{
A list() object with oucome and covariates splitted
}
\description{
This function splits the passed formula into outcome and covariates
}
