% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stabilise_levels.R
\name{stabilise_levels}
\alias{stabilise_levels}
\title{Multiply each sample IPTW for their specific treatment group percentage on the whole cohort}
\usage{
stabilise_levels(data, outcome, outcome_levels = NULL, iptw = "iptw")
}
\arguments{
\item{data:}{the data.frame object for the analysis, output of add_IPTW()}

\item{outcome:}{the column name of the treatment variable}

\item{outcome_levels:}{the levels of the treatment variable (optional: levels() of the column, it will be converted to factor() in case is necessary)}

\item{iptw:}{the column name of the IPTW variable (output of add_IPTW())}
}
\value{
A vector object where IPTW is corrected for the percentage of cases in that specific tratment group for each sample
}
\description{
This function multiply each sample IPTW value (as per add_IPTW()) for
the percentage of cases on that specific treatment variable group
}
